RDMA engine server mode flow fix ups.
When running fio in RDMA server mode, the user just sees an ever
increasing ETA line. Which is rather uninfromative. This patch
adds a waiting for connection message and sets the thread to SETTING_UP
while waiting.

Then when the server is running the ETA is disabled. (The RDMA server
has no access to progress information.)